// src/interfaces/IClient.ts
export interface IClient {
  id?: number;
  active: boolean;
  address: string;
  addressNumber: string;
  apelido: string;
  cei: string;
  cel: string;
  celResponsible: string;
  city: string;
  cityResponsible: string;
  cnh: string;
  complement: string;
  contribuinte: string;
  dataOfBirthdate: string;
  dataOfBirthdateResponsible: string;
  document: string;
  documentResponsible: string;
  email: string;
  emailResponsible: string;
  issuer: string;
  maritalStatus: string;
  name: string;
  nameResponsible: string;
  neighborhood: string;
  observation: string;
  personType: string;
  registerCity: string;
  registerDocument: string;
  registerState: string;
  security: string;
  tel: string;
  telResponsible: string;
  uf: string;
  ufDocument: string;
  zipCode: string;
  zipCodeResponsible: string;
}